+++ ts-sdk/api/tests/markets.spec.ts
@@ -1,17 +1,14 @@
 import { describe, expect, it } from "vitest";
 
-import { getMarket, getMarkets, setTunaBaseUrl, unwrap } from "../src";
+import { unwrap } from "../src";
 
 import { SOL_USDC_FUSION_POOL_ADDRESS } from "./consts";
 import { getMarketsFromRpc } from "./rpc";
-
-import "dotenv/config";
-
-setTunaBaseUrl(process.env.API_BASE_URL!);
+import { sdk } from "./sdk";
 
 describe("Markets", async () => {
   const rpcMarkets = await getMarketsFromRpc();
-  const markets = await unwrap(getMarkets());
+  const markets = await unwrap(sdk.getMarkets());
 
   it("Length matches rpc", () => {
     expect(markets.length).toBe(rpcMarkets.length);
@@ -32,7 +29,11 @@ describe("Single Market", async () => {
   const marketDst = rpcMarkets.find(market => market.data.pool === SOL_USDC_FUSION_POOL_ADDRESS) ?? rpcMarkets[0];
   const sampleMarketAddress = marketDst.address;
   const unsavedMarketAddress = "FeR8VBqNRSUD5NtXAj2n3j1dAHkZHfyDktKuLXD4pump";
-  const market = await unwrap(getMarket(sampleMarketAddress));
+  const market = await unwrap(
+    sdk.getMarket({
+      marketAddress: sampleMarketAddress,
+    }),
+  );
 
   it("Returns market data", () => {
     expect(market.address).toBe(sampleMarketAddress);
@@ -40,11 +41,15 @@ describe("Single Market", async () => {
   });
 
   it("Returns 404 for unsaved market", async () => {
-    const response = await getMarket(unsavedMarketAddress);
+    const { response } = await sdk.getMarket({
+      marketAddress: unsavedMarketAddress,
+    });
     expect(response.status).toBe(404);
   });
   it("Returns 400 for invalid market", async () => {
-    const response = await getMarket("123");
+    const { response } = await sdk.getMarket({
+      marketAddress: "123",
+    });
     expect(response.status).toBe(400);
   });
 });